[web] Detach comment-related code from home
